perf: don't use private members
Because I am running this on Node 10, which doesn't have real private
members, I'm using a "target" of `ES2019` in the tsconfig.json, and that
means that TypeScript is doing a fair bit of fancy abstraction during
compilation, and you can see that in this subset of the profiler output:

   ticks  total  nonlib   name
     62    7.1%    7.2%  Builtin: WeakMapLookupHashIndex
     37    4.2%    4.3%  Builtin: WeakMapGet
     23    2.6%    2.7%  Builtin: WeakMapHas
     21    2.4%    2.4%  LazyCompile: *__classPrivateFieldGet /Users/glh/code/algorithms/union-find/percolate/src/UnionFind.js:9:80
     12    1.4%    1.4%  LazyCompile: *__classPrivateFieldGet /Users/glh/code/algorithms/union-find/percolate/src/Percolation.js:9:80

So that's (at least) 17.7% of all samples landing in something related
to private members (and the WeakMaps that are used in their
implementation). That profile run produced with:

    VERBOSE=0 node --prof src/montecarlo.js 1000 1
    node --prof-process isolate-0x103001e00-v8.log > processed.txt

Let's rip out the private members and write them as
"private-by-convention-only" public members (ie. with an underscore
prefix).

    Before (best of three runs):

        VERBOSE=0 yarn go 1000 1
        Runs: 1
        Size: 1000 * 1000
        Total time: 845.17ms (845.17ms per run)
        Threshhold: 0.593627

    After (best of three runs):

        VERBOSE=0 yarn go 1000 1
        Runs: 1
        Size: 1000 * 1000
        Total time: 340.69ms (340.69ms per run)
        Threshhold: 0.596138