test: show that percolation detection is probably right
When I wrote it, I was a bit worried about sensitivity to ordering
issues, because the order in which we open cells determines how leaders
are chosen over time, and how ranks change etc, all of which could
invalidate our assumption that we can reliably look up the "lowest"
entry for a given leader.

This test isn't exhaustive (in the sense that it only opens cells in the
first column), but it it does test all permutations (tens of thousands
of them) in for an 8x8 grid, so it can give us some confidence that the
solution probably works.