perf: avoid temporary array
Simplify minimum/maximum detection by sorting on slope then tiebreaking
based on natural ordering. And then use a bit of book-keeping to check
for beginning/end of runs without actually materializing them in the
`equal` array. Note that we preserve the trick from the previous commit
of avoiding duplicates by only pushing when `p` is the smallest point in
the segment.