chore: provide sample TS files for 8-puzzle exercise
By overwriting all the files from the previous week's exercise, which
is why you see a bunch of deletions in addition to the additions this
commit.

The heap implementation in `Heap.ts` can be used to construct a priority
queue. A sample such queue can be found in `src/MinPQ.ts`.