chore: prepare to work on kd-trees exercise
This takes our 8-puzzle-solution and strips out all of its sample files,
replacing them with the new sample files for this week.

I deleted most of the code, but kept the heap and min-priority queue
implementations because I expect I may need them, based on my skimming
of the problem specification.