Revert "perf: swap out LSD radix sort for 3-way radix quicksort"
As planned, this reverts commit 1a6047ebbf88e.

We could implement a fancy cut-over system where we use 3-way radix
quicksort for short strings, and fallback to LSD radix sort for longer
strings, but the exact cut-off point probably varies according to the
input string (because the structure of the recursive calls will change),
so it seems error prone to try and pick a point that is guaranteed to
work.